+++ ricaun.DA4R.NUnit/Services/DesignAutomationController.cs
@@ -1,6 +1,7 @@
 ﻿using Autodesk.Revit.ApplicationServices;
 using Autodesk.Revit.DB;
 using ricaun.DA4R.NUnit.Models;
+using ricaun.DA4R.NUnit.Revit;
 using System;
 using System.Diagnostics;
 using System.Linq;
@@ -9,18 +10,14 @@ using System.Runtime.Versioning;
 
 namespace ricaun.DA4R.NUnit.Services
 {
-    public class DesignAutomationController
+    public class DesignAutomationController : IDesignAutomation
     {
-        public static bool Execute(Application revitApp)
-        {
-            return Execute(revitApp, null, null);
-        }
-
-        public static bool Execute(Application application, string filePath, Document revitDoc)
+        public bool Execute(Application application, string filePath = null, Document revitDoc = null)
         {
             RevitParameters.AddParameter(application, application.GetControlledApplication());
 
             Console.WriteLine($"Execute: {typeof(DesignAutomationController).Assembly.GetCustomAttribute<TargetFrameworkAttribute>()?.FrameworkName}");
+            Console.WriteLine($"TestEngine: {typeof(ricaun.NUnit.TestEngine).Assembly.FullName}");
 
             var output = new OutputModel();